move creation of nats subnet to nats class
