+++ aihub_iac/aihub_iac/azure/modules/nats/Nats.py
@@ -1,9 +1,11 @@
 from typing import Optional
 
 import pulumi
-from pulumi_azure_native import containerinstance
+from pulumi_azure_native import containerinstance, network
 
+from aihub_iac.azure.constants.resources import SUB_NET, STORAGE_ACCOUNT, CONTAINER_INSTANCE
 from aihub_iac.azure.modules.nats.NatsConfig import NatsConfig
+from aihub_iac.azure.modules.network.Network import APP_SUBNET_PREFIX, AGENTS_SUBNET_PREFIX, DAGSTER_SUBNET_PREFIX
 from aihub_iac.azure.providers.NetworkProvider import NetworkProvider
 from aihub_iac.azure.resources.storage.StorageResourceFactory import StorageResourceFactory
 
@@ -16,6 +18,9 @@ class Nats(pulumi.ComponentResource):
     Nats(stack, name, config=config)
     """
 
+    NATS_SUBNET_CIDR = "10.0.1.0/29"
+    NATS_STORAGE_SUBNET_CIDR = "10.0.32.0/24"
+
     def __init__(self, stack: str, name: str, config: NatsConfig, opts: Optional[pulumi.ResourceOptions] = None):
         super().__init__(f"{stack}:{name}", name, None, opts)
 
@@ -27,6 +32,14 @@ class Nats(pulumi.ComponentResource):
             self.config.resource_group, self.config.project_name, self.config.location_short
         )
         self._create_resources()
+
+    @property
+    def nats_storage_subnet_name(self):
+        return f"{self.config.project_name}-{SUB_NET}-{self.config.location_short}-{STORAGE_ACCOUNT}-nats"
+
+    @property
+    def nats_subnet_name(self):
+        return f"{self.config.project_name}-{SUB_NET}-{self.config.location_short}-{CONTAINER_INSTANCE}-nats"
 
     def _create_nats_container(self) -> containerinstance.ContainerArgs:
         """Create a NATS container configuration"""
@@ -62,17 +75,59 @@ class Nats(pulumi.ComponentResource):
             ),
         )
 
+    def _create_nats_subnet(self) -> network.Subnet:
+        subnet = network.Subnet(
+            name=self.nats_subnet_name,
+            resource_name=self.nats_subnet_name,
+            resource_group_name=self.config.resource_group,
+            virtual_network_name=self.vnet.name,
+            address_prefix=self.NATS_SUBNET_CIDR,
+            delegations=[
+                network.DelegationArgs(
+                    name="aci-delegation",
+                    service_name="Microsoft.ContainerInstance/containerGroups",
+                )
+            ],
+            opts=pulumi.ResourceOptions(parent=self.vnet),
+        )
+        self.network_provider.create_subnet_nsg(
+            self.nats_subnet_name,
+            subnet,
+            [APP_SUBNET_PREFIX, AGENTS_SUBNET_PREFIX, DAGSTER_SUBNET_PREFIX],
+        )
+        return subnet
+
+    def _create_nats_storage_subnet(self) -> network.Subnet:
+        subnet = network.Subnet(
+            name=self.nats_storage_subnet_name,
+            resource_name=self.nats_storage_subnet_name,
+            resource_group_name=self.config.resource_group,
+            virtual_network_name=self.vnet.name,
+            address_prefix=self.NATS_STORAGE_SUBNET_CIDR,
+            opts=pulumi.ResourceOptions(parent=self.vnet),
+        )
+
+        self.network_provider.create_subnet_nsg(
+            self.nats_storage_subnet_name,
+            subnet,
+            [self.NATS_SUBNET_CIDR, self.NATS_STORAGE_SUBNET_CIDR],
+        )
+        return subnet
+
     def _create_resources(self):
         """Create all required resources with proper dependency management"""
         self.vnet = self.network_provider.get_vnet()
-        self.nats_storage_subnet = self.network_provider.get_nats_storage_subnet()
-        self.nats_subnet = self.network_provider.get_nats_subnet()
+        self.nats_storage_subnet = self._create_nats_storage_subnet()
+        self.nats_subnet = self._create_nats_subnet()
 
-        self.storage_account = self.storage_factory.create_storage_account(
-            service_name=self.config.storage_service_name,
-            subnet_id=self.nats_storage_subnet.id,
-            vnet_id=self.vnet.id,
+        self.storage_account = pulumi.Output.all(subnet_id=self.nats_storage_subnet.id, vnet_id=self.vnet.id).apply(
+            lambda args: self.storage_factory.create_storage_account(
+                service_name=self.config.storage_service_name,
+                subnet_id=args["subnet_id"],
+                vnet_id=args["vnet_id"],
+            )
         )
+
         self.storage_account_key = self.storage_factory.get_storage_account_key(self.storage_account)
 
         self.nats_file_share = self.storage_factory.create_file_share("blob", self.storage_account)
