+++ aihub_iac/aihub_iac/azure/providers/NetworkProvider.py
@@ -1,3 +1,4 @@
+import pulumi
 from pulumi_azure_native import network
 
 from aihub_iac.azure.constants.resources import (
@@ -19,6 +20,79 @@ class NetworkProvider:
         self.project_name = project_name
         self.location_short_name = location_short_name
 
+    def create_subnet_nsg(
+        self,
+        subnet_name: str,
+        subnet: network.Subnet,
+        source_prefixes: list[str],
+        additional_rules: list[network.SecurityRuleArgs] = None,
+    ) -> network.NetworkSecurityGroup:
+        """Create NSG for a subnet that only allows traffic from specific source subnets
+
+        Args:
+            subnet: The subnet to create NSG for
+            source_prefixes: List of source address prefixes to allow traffic from
+        """
+        # Create security rules for each source prefix
+        security_rules = []
+        for idx, prefix in enumerate(source_prefixes):
+            security_rules.append(
+                network.SecurityRuleArgs(
+                    name=f"AllowFromSourceSubnet{idx+1}",
+                    priority=100 + idx,  # Increment priority for each rule
+                    direction="Inbound",
+                    access="Allow",
+                    protocol="*",
+                    source_address_prefix=prefix,
+                    source_port_range="*",
+                    destination_address_prefix="*",
+                    destination_port_range="*",
+                )
+            )
+
+        # Add any additional rules provided
+        if additional_rules:
+            # Start priorities after the source subnet rules
+            start_priority = 100 + len(source_prefixes)
+            for idx, rule in enumerate(additional_rules):
+                # If the rule doesn't have a priority set, assign one
+                if not hasattr(rule, "priority") or rule.priority is None:
+                    rule.priority = start_priority + idx
+                security_rules.append(rule)
+
+        # Add the deny rule at the end
+        security_rules.append(
+            network.SecurityRuleArgs(
+                name="DenyAllInbound",
+                priority=4096,
+                direction="Inbound",
+                access="Deny",
+                protocol="*",
+                source_address_prefix="*",
+                source_port_range="*",
+                destination_address_prefix="*",
+                destination_port_range="*",
+            )
+        )
+
+        nsg = network.NetworkSecurityGroup(
+            resource_name=f"{subnet_name}-nsg",
+            network_security_group_name=f"{subnet_name}-nsg",
+            resource_group_name=self.config.resource_group,
+            location=self.config.location,
+            security_rules=security_rules,
+            id=subnet.id,
+            opts=pulumi.ResourceOptions(
+                parent=self,
+                replace_on_changes=["security_rules"],
+            ),
+            tags={
+                "Stack": self.stack,
+            },
+        )
+
+        return nsg
+
     def get_subnet(self, subnet_name=None):
         return network.get_subnet_output(
             resource_group_name=self.resource_group,
@@ -32,9 +106,6 @@ class NetworkProvider:
             virtual_network_name=self.v_net_name,
         )
 
-    def get_nats_subnet(self):
-        return self.get_subnet(subnet_name=self.nats_subnet_name)
-
     def get_cap_subnet(self):
         return self.get_subnet(subnet_name=self.cap_subnet_name)
 
@@ -46,9 +117,6 @@ class NetworkProvider:
 
     def get_phoenix_subnet(self):
         return self.get_subnet(subnet_name=self.phoenix_subnet_name)
-
-    def get_nats_storage_subnet(self):
-        return self.get_subnet(subnet_name=self.nats_storage_subnet_name)
 
     def get_webui_storage_subnet(self):
         return self.get_subnet(subnet_name=self.webui_storage_subnet_name)
@@ -111,10 +179,6 @@ class NetworkProvider:
         return f"{self.sub_net_name}-{CONTAINER_INSTANCE}-agents"
 
     @property
-    def nats_storage_subnet_name(self):
-        return f"{self.sub_net_name}-{STORAGE_ACCOUNT}-nats"
-
-    @property
     def webui_storage_subnet_name(self):
         return f"{self.sub_net_name}-{STORAGE_ACCOUNT}-webui"
 
@@ -134,9 +198,6 @@ class NetworkProvider:
     def priv_endpoint_subnet_name(self):
         return f"{self.sub_net_name}-priv-endpoint"
 
-    @property
-    def nats_subnet_name(self):
-        return f"{self.sub_net_name}-{CONTAINER_INSTANCE}-nats"
 
     @property
     def dagster_storage_subnet_name(self):
